Move access_url_first_filtered to application_helper so exporters can see it
